# Speed up incremental package asset resolution