--- a/src/Tasks/Microsoft.NET.Build.Tasks/ReportAssetsLogMessages.cs
+++ b/src/Tasks/Microsoft.NET.Build.Tasks/ReportAssetsLogMessages.cs
@@ -9,9 +9,9 @@ using System.Linq;
 namespace Microsoft.NET.Build.Tasks
 {
     /// <summary>
-    /// Report Log Messages in the assets file to MSBuild and raise them as
-    /// DiagnosticMessage items that can be consumed downstream (e.g. by the
-    /// dependency node in the solution explorer)
+    /// Raise log messages in the assets file as DiagnosticMessage items
+    /// that can be consumed downstream (e.g. by the dependency node in
+    /// the solution explorer)
     /// </summary>
     public sealed class ReportAssetsLogMessages : TaskBase
     {
@@ -40,15 +40,14 @@ namespace Microsoft.NET.Build.Tasks
 
         public ReportAssetsLogMessages()
         {
-           _diagnostics = new DiagnosticsHelper(new MSBuildLog(Log));
+           _diagnostics = new DiagnosticsHelper();
         }
 
         #region Test Support
 
-        internal ReportAssetsLogMessages(LockFile lockFile, ILog logger)
+        internal ReportAssetsLogMessages(LockFile lockFile) : this()
         {
             _lockFile = lockFile;
-            _diagnostics = new DiagnosticsHelper(logger);
         }
 
         #endregion