… src/Tests/Microsoft.NET.Build.Tests/GivenThatWeWantBuildsToBeIncremental.cs
… src/Tests/Microsoft.NET.Build.Tests/GivenThatWeWantToVerifyPCLProjectReferenceCompat.cs
… src/Tests/Microsoft.NET.Restore.Tests/GivenThatWeWantToRestoreProjectsWithPackageDowngrades.cs
… src/Tests/Microsoft.NET.TestFramework/Commands/MSBuildCommand.cs