Whenever project.assets.json changes (or more rarely, a setting influencing ResolvePackageAssets changes), write out a binary cache file for the next incremental build. The file has only the items that used by the build and they are written in a format that can be deserialized to MSBuild ITaskItem[] with minimal time and allocation. 

On top of that, we merge ReportAssetsLogMessages into ResolvePackageAssets, which allows for project.assets.json to not be read at all if it has not changed.

Some early measurement below using https://github.com/mikeharder/dotnet-cli-perf/tree/master/scenarios/web/large/core

First observation, the total size of all assets.cache files is 5 MB vs. 37 MB for assets.json.

I will send a private drop to our internal perf alias as wll.

## Incremental build --no-restore /m:1
```
Before 
      161 ms  JoinItems                                129 calls
      176 ms  AssignProjectConfiguration               129 calls
      264 ms  Copy                                     370 calls
      318 ms  ResolvePackageFileConflicts              129 calls
     1180 ms  ReportAssetsLogMessages                  129 calls
     2873 ms  ResolvePackageAssets                     129 calls
    23138 ms  ResolveAssemblyReference                 129 calls
    
    Time Elapsed 00:00:40.08
    
After
      117 ms  JoinItems                                129 calls
      127 ms  GetReferenceNearestTargetFrameworkTask   109 calls
      136 ms  AssignProjectConfiguration               129 calls
      150 ms  Message                                  388 calls
      252 ms  ResolvePackageAssets                     129 calls
      257 ms  ResolvePackageFileConflicts              129 calls
      307 ms  Copy                                     370 calls
    23261 ms  ResolveAssemblyReference                 129 calls

    Time Elapsed 00:00:36.04
```

Saving 10% overall here and reducing ResolvePackageAssets+ReportAssetsLogMessage (now combined into ResolvePackageAssets) from ~4 seconds to ~250 milliseconds.

## Incremental parallel build --no-restore
```
Before
      184 ms  Hash                                     258 calls
      232 ms  JoinItems                                129 calls
      240 ms  GetReferenceNearestTargetFrameworkTask   109 calls
      299 ms  AssignProjectConfiguration               129 calls
      487 ms  ResolvePackageFileConflicts              129 calls
      492 ms  Copy                                     370 calls
     2654 ms  ReportAssetsLogMessages                  129 calls
     5071 ms  ResolvePackageAssets                     129 calls
    37575 ms  ResolveAssemblyReference                 129 calls
    
    Time Elapsed 00:00:19.72

After
      100 ms  ConvertToAbsolutePath                    129 calls
      116 ms  ProduceContentAssets                     129 calls
      157 ms  FindUnderPath                            645 calls
      185 ms  Hash                                     258 calls
      246 ms  JoinItems                                129 calls
      329 ms  AssignProjectConfiguration               129 calls
      447 ms  GetReferenceNearestTargetFrameworkTask   109 calls
      487 ms  ResolvePackageFileConflicts              129 calls
      502 ms  Copy                                     370 calls
      612 ms  ResolvePackageAssets                     129 calls
    38504 ms  ResolveAssemblyReference                 129 calls

    Time Elapsed 00:00:18.54
```
Overall saving is less in parallel case, but time in   (ResolvePackageAssets + ReportAssetsLogMessage)  goes down from ~7.5 seconds to ~ 600 milliseconds.

It is interesting that just about every task is taking significantly more cumulative time when building in parallel. But, of course, some (most???) of the 7 seconds difference is being recuperated by parallelism.


## Clean parallel build --no-restore (cleaned and restored beforehand)
```
Before
      168 ms  Hash                                     258 calls
      270 ms  GetReferenceNearestTargetFrameworkTask   109 calls
      301 ms  JoinItems                                129 calls
      306 ms  MakeDir                                  129 calls
      333 ms  AssignProjectConfiguration               129 calls
      524 ms  ResolvePackageFileConflicts              129 calls
      559 ms  WriteCodeFragment                        129 calls
     1866 ms  WriteLinesToFile                         387 calls
     3268 ms  ReportAssetsLogMessages                  129 calls
     4616 ms  GenerateDepsFile                         129 calls
     5577 ms  ResolvePackageAssets                     129 calls
     6945 ms  Copy                                     370 calls
    24787 ms  Csc                                      129 calls
    41201 ms  ResolveAssemblyReference                 129 calls
    
    Time Elapsed 00:00:29.46
    
After
      167 ms  Hash                                     258 calls
      181 ms  FindUnderPath                            645 calls
      234 ms  GetReferenceNearestTargetFrameworkTask   109 calls
      270 ms  AssignProjectConfiguration               129 calls
      294 ms  JoinItems                                129 calls
      299 ms  MakeDir                                  129 calls
      487 ms  ResolvePackageFileConflicts              129 calls
      540 ms  WriteCodeFragment                        129 calls
     1501 ms  WriteLinesToFile                         387 calls
     4317 ms  GenerateDepsFile                         129 calls
     6802 ms  Copy                                     370 calls
     9715 ms  ResolvePackageAssets                     129 calls
    26625 ms  Csc                                      129 calls
    39653 ms  ResolveAssemblyReference                 129 calls
    
    Time Elapsed 00:00:30.44
```
This case shows the *slowdown* of  (ResolvePackageAssets + ReportAssetsLogMessage)  from ~8.8 seconds to ~9.7 seconds in the first run of ResolvePackageAssets where the cache needs to be generated. I have some ideas on how to narrow this down significantly, but note that it does not impact the inner loop case where project.assets.json would not be changing between every build.

cc @mikeharder @davkean @livarcocc @dsplaisted