@mikeharder 2018-03-06
Here are my results for WebLargeCore:

# Incremental /m:1
Version | Time (s)
--- | --- 
2.1.300-preview1 (8174) | TBD
2.1.300-preview2-008293 | TBD
2.1.300-preview2-008293-ResolvePackageAssets | TBD

# Incremental Parallel
Version | Time (s)
--- | --- 
2.1.300-preview1 (8174) | 24.3
2.1.300-preview2-008293 | 22.6
2.1.300-preview2-008293-ResolvePackageAssets | 22.1