@mikeharder 2018-03-06
My results are similar to @nguerrera:

Parallel|Core @ 8293 (s)|Core @ 8293 + ResolvePackageAssets (s)|Delta (s)|Delta %
-----|-----|-----|-----|-----
1|52.8|48.8|4.0|7.6%
8|22.6|22.1|0.5|2.0% 

The improvement is significantly larger for a single-proc build, even on a relative basis.