@nguerrera 2018-03-06
The symptom that every task is slower in parallel builds is reminiscent of the logging bug. We should check that it hasn't regressed. It's very strange to me that all that cumulative time in ResolvePackageAssets (and ReportAssetsLogMessages) is gone but only such a small fraction of those savings accrue to elapsed time.