@mikeharder 2018-03-06
I agree we should do more investigation of the relative performance between `/m:1` and `/m:8` (for an 8-core machine).  The solution has 129 projects, and most of the "layers" in the dependency graph have at least 8 projects, so I would expect it to parallelize quite well.  However, `/m:8` is only about 2x faster than `/m:1`, compared to a theoretical maximum of 8x faster.

We should look at things like:

* How is the work scheduled across the 8 nodes?  Is a lot of time spent waiting?
* What is the overall CPU utilization?
* What is the cost of sending data between nodes (including serialization)?

CC: @rainersigwald