@rainersigwald 2018-03-06
Took a quick look at the /detailedsummary output of an /m:8 no-op incremental build on sdk 2.1.100 (msbuild 15.6.82.30579)

```
Utilization:          37.9    47.2    86.3    43.0    31.7    83.1    84.3    82.3     Average Utilization: 62.0
```

That's definitely not great, but I'd expect it to do better than a 2x speedup.

Looks like the node builds are densely packed for the first 44 seconds of the build and only then does the utilization graph start to have holes.

Definitely deserves a deeper look.