@nguerrera 2018-03-15
> It seems like it could be a good idea to have unit-level tests for reading and writing the cache file.

It certainly wouldn't hurt, but as noted elsewhere every single integration test in the suite is going to be writing/reading from the cache. In fact, it is designed to go through the same path in full builds as incremental builds for exactly that reason. I noted in #2048 that if we take a different approach we will need much more testing for the incremental case.