@nguerrera 2018-03-15
@dotnet-bot Test OSX10.12 Debug please