@nguerrera 2018-03-16
@dotnet-bot Test OSX10.12 Debug 

Filed #2052 

The flakiness predates this change. I'm disabling the test on OS X if it fails again.